+++ src/sys/sos_main.c
@@ -1,13 +1,13 @@
 // Copyright 2011-2021 Tyler Gilbert and Stratify Labs, Inc; see LICENSE.md
 
+#include <sdk/api.h>
+
 #include "cortexm/cortexm.h"
 
 #include "check_config.h"
 
 #include "scheduler/scheduler_timing.h"
 
-static void check_sys_config();
-static void check_debug_config();
 static void check_config();
 
 int sos_main() {
@@ -22,20 +22,37 @@ int sos_main() {
 #endif
     sos_config.debug.initialize();
 
-    sos_debug_log_directive(SOS_DEBUG_MALLOC | SOS_DEBUG_TASK | SOS_DEBUG_SCHEDULER, "reset");
-    sos_debug_log_directive(SOS_DEBUG_MALLOC, "hist:Malloc Perf:_malloc_r_us:malloc() execution time in us");
-    sos_debug_log_directive(SOS_DEBUG_MALLOC, "hist:Free Perf:_free_r_us:free() execution time in us");
-    sos_debug_log_directive(SOS_DEBUG_PTHREAD, "hist:Mutex Lock Perf:pthread_mutex_lock_us:pthread_mutex_lock() execution time in us");
-    sos_debug_log_directive(SOS_DEBUG_PTHREAD, "hist:Mutex Unlock Perf:pthread_mutex_unlock_us:pthread_mutex_unlock() execution time in us");
-    sos_debug_log_directive(SOS_DEBUG_PTHREAD, "hist:Condition Perf:pthread_cond_timedwait_us:pthread_cond_timedwait() execution time in us");
-    sos_debug_log_directive(SOS_DEBUG_UNISTD, "hist:usleep Oversleep:usleep_us:usleep() oversleep time in us");
-    sos_debug_log_directive(SOS_DEBUG_UNISTD, "hist:sleep Oversleep:sleep_us:sleep() oversleep time in us");
-    sos_debug_log_directive(SOS_DEBUG_TASK, "hist:Context Switch Cycles:switch_contexts:Average number of cycles for context switching");
-    sos_debug_log_directive(SOS_DEBUG_TASK, "hist:PendSV Critical Cycles:pend_critical:Average number of cycles interrupts are off on PendSV");
-    sos_debug_log_directive(SOS_DEBUG_SCHEDULER, "hist:Scheduler Critical Cycles:scheduler_critical:Average number of cycles interrupts are off when stopping a task");
-    sos_debug_log_directive(SOS_DEBUG_MALLOC, "heap:OS Heap:heap0:OS Heap Utilization over time");
+    sos_debug_log_directive(
+      SOS_DEBUG_MALLOC | SOS_DEBUG_TASK | SOS_DEBUG_SCHEDULER, "reset");
+    sos_debug_log_directive(
+      SOS_DEBUG_MALLOC, "hist:Malloc Perf:_malloc_r_us:malloc() execution time in us");
+    sos_debug_log_directive(
+      SOS_DEBUG_MALLOC, "hist:Free Perf:_free_r_us:free() execution time in us");
+    sos_debug_log_directive(
+      SOS_DEBUG_PTHREAD,
+      "hist:Mutex Lock Perf:pthread_mutex_lock_us:pthread_mutex_lock() execution time in us");
+    sos_debug_log_directive(
+      SOS_DEBUG_PTHREAD,
+      "hist:Mutex Unlock Perf:pthread_mutex_unlock_us:pthread_mutex_unlock() execution time in us");
+    sos_debug_log_directive(
+      SOS_DEBUG_PTHREAD,
+      "hist:Condition Perf:pthread_cond_timedwait_us:pthread_cond_timedwait() execution time in us");
+    sos_debug_log_directive(
+      SOS_DEBUG_UNISTD, "hist:usleep Oversleep:usleep_us:usleep() oversleep time in us");
+    sos_debug_log_directive(
+      SOS_DEBUG_UNISTD, "hist:sleep Oversleep:sleep_us:sleep() oversleep time in us");
+    sos_debug_log_directive(
+      SOS_DEBUG_TASK,
+      "hist:Context Switch Cycles:switch_contexts:Average number of cycles for context switching");
+    sos_debug_log_directive(
+      SOS_DEBUG_TASK,
+      "hist:PendSV Critical Cycles:pend_critical:Average number of cycles interrupts are off on PendSV");
+    sos_debug_log_directive(
+      SOS_DEBUG_SCHEDULER,
+      "hist:Scheduler Critical Cycles:scheduler_critical:Average number of cycles interrupts are off when stopping a task");
+    sos_debug_log_directive(
+      SOS_DEBUG_MALLOC, "heap:OS Heap:heap0:OS Heap Utilization over time");
   }
-
 
   check_config();
 
@@ -89,68 +106,80 @@ bootloader_api_t *sos_get_bootloader_api() {
   return NULL;
 }
 
-void check_sys_config() {
-  // SOS_EVENT_ROOT_PANIC_SYSTEM_CONFIG
-}
-
-void check_debug_config() {
-  // SOS_EVENT_ROOT_PANIC_DEBUG_CONFIG
-}
-
 void check_config() {
-  CHECK_SYS_CONFIG(get_serial_number);
-  CHECK_SYS_CONFIG(kernel_request);
-  CHECK_SYS_CONFIG(kernel_request_api);
-  CHECK_SYS_CONFIG(name);
-  CHECK_SYS_CONFIG(version);
-  CHECK_SYS_CONFIG(git_hash);
-  CHECK_SYS_CONFIG(id);
-  CHECK_SYS_CONFIG(team_id);
+  CHECK_CONFIG(sys, get_serial_number);
+  CHECK_CONFIG(sys, kernel_request);
+  CHECK_CONFIG(sys, kernel_request_api);
+  CHECK_CONFIG(sys, name);
+  CHECK_CONFIG(sys, version);
+  CHECK_CONFIG(sys, git_hash);
+  CHECK_CONFIG(sys, id);
+  CHECK_CONFIG(sys, team_id);
+  CONFIG_ASSERT(
+    sos_config.sys.memory_size > 8192, "assert: sos_config.sys.memory_size > 8192")
+  CHECK_CONFIG(sys, initialize);
 
-  CHECK_MCU_CONFIG(set_interrupt_priority);
-  CHECK_MCU_CONFIG(reset_watchdog_timer);
-  CHECK_MCU_CONFIG(set_pin_function);
+  CHECK_CONFIG(mcu, set_interrupt_priority);
+  CHECK_CONFIG(mcu, reset_watchdog_timer);
+  CHECK_CONFIG(mcu, set_pin_function);
 
-  CHECK_SLEEP_CONFIG(idle);
-  CHECK_SLEEP_CONFIG(hibernate);
-  CHECK_SLEEP_CONFIG(powerdown);
+  CHECK_CONFIG(fs, devfs_list);
+  CHECK_CONFIG(fs, rootfs_list);
+  CHECK_CONFIG(fs, stdin_dev);
+  CHECK_CONFIG(fs, stdout_dev);
+  CHECK_CONFIG(fs, stderr_dev);
+  CHECK_CONFIG(fs, trace_dev);
 
-  CHECK_CACHE_CONFIG(enable);
-  CHECK_CACHE_CONFIG(disable);
-  CHECK_CACHE_CONFIG(invalidate_instruction);
-  CHECK_CACHE_CONFIG(invalidate_data);
-  CHECK_CACHE_CONFIG(invalidate_data_block);
-  CHECK_CACHE_CONFIG(clean_data);
-  CHECK_CACHE_CONFIG(clean_data_block);
+  CHECK_CONFIG(sleep, idle);
+  CHECK_CONFIG(sleep, hibernate);
+  CHECK_CONFIG(sleep, powerdown);
 
-  CHECK_CACHE_POLICY_CONFIG(external_sram_policy);
-  CHECK_CACHE_POLICY_CONFIG(sram_policy);
-  CHECK_CACHE_POLICY_CONFIG(flash_policy);
-  CHECK_CACHE_POLICY_CONFIG(peripherals_policy);
-  CHECK_CACHE_POLICY_CONFIG(lcd_policy);
-  CHECK_CACHE_POLICY_CONFIG(external_flash_policy);
-  CHECK_CACHE_POLICY_CONFIG(tightly_coupled_data_policy);
-  CHECK_CACHE_POLICY_CONFIG(tightly_coupled_instruction_policy);
-  CHECK_CACHE_POLICY_CONFIG(os_data_mpu_type);
-  CHECK_CACHE_POLICY_CONFIG(os_system_data_mpu_type);
+  CHECK_CONFIG(debug, enable_led);
+  CHECK_CONFIG(debug, disable_led);
+
+  CHECK_CONFIG(cache, enable);
+  CHECK_CONFIG(cache, disable);
+  CHECK_CONFIG(cache, invalidate_instruction);
+  CHECK_CONFIG(cache, invalidate_data);
+  CHECK_CONFIG(cache, invalidate_data_block);
+  CHECK_CONFIG(cache, clean_data);
+  CHECK_CONFIG(cache, clean_data_block);
+
+  CHECK_CONFIG(cache, external_sram_policy);
+  CHECK_CONFIG(cache, sram_policy);
+  CHECK_CONFIG(cache, flash_policy);
+  CHECK_CONFIG(cache, peripherals_policy);
+  CHECK_CONFIG(cache, lcd_policy);
+  CHECK_CONFIG(cache, external_flash_policy);
+  CHECK_CONFIG(cache, tightly_coupled_data_policy);
+  CHECK_CONFIG(cache, tightly_coupled_instruction_policy);
+  CHECK_CONFIG(cache, os_code_mpu_type);
+  CHECK_CONFIG(cache, os_data_mpu_type);
+  CHECK_CONFIG(cache, os_system_data_mpu_type);
 
   // cache policy check
   if (sos_config.usb.logical_endpoint_count > 0) {
-    CHECK_USB_CONFIG(set_attributes);
-    CHECK_USB_CONFIG(set_action);
-    CHECK_USB_CONFIG(write_endpoint);
-    CHECK_USB_CONFIG(read_endpoint);
+    CONFIG_ASSERT(
+      sos_config.usb.control_endpoint_max_size != 0,
+      "assert: sos_config.usb.control_endpoint_max_size != 0");
+
+    CHECK_CONFIG(usb, set_attributes);
+    CHECK_CONFIG(usb, set_action);
+    CHECK_CONFIG(usb, write_endpoint);
+    CHECK_CONFIG(usb, read_endpoint);
   }
 
+  CHECK_CONFIG(task, start);
+  CONFIG_ASSERT(
+    sos_config.task.start_stack_size > CONFIG_PTHREAD_STACK_MIN,
+    "assert: sos_config.task.start_stack_size > CONFIG_PTHREAD_STACK_MIN")
 
-  CHECK_TASK_CONFIG(start);
-
-  CHECK_CLOCK_CONFIG(initialize);
-  CHECK_CLOCK_CONFIG(enable);
-  CHECK_CLOCK_CONFIG(disable);
-  CHECK_CLOCK_CONFIG(set_channel);
-  CHECK_CLOCK_CONFIG(get_channel);
-  CHECK_CLOCK_CONFIG(microseconds);
-  CHECK_CLOCK_CONFIG(nanoseconds);
+  CHECK_CONFIG(clock, initialize);
+  CHECK_CONFIG(clock, enable);
+  CHECK_CONFIG(clock, disable);
+  CHECK_CONFIG(clock, set_channel);
+  CHECK_CONFIG(clock, get_channel);
+  CHECK_CONFIG(clock, microseconds);
+  CHECK_CONFIG(clock, nanoseconds);
 
 }